My jijitang page completed, code clean and well organized. my jijitang page is composed of three files, Mypage.html, MyEntries.html and myEntries.js.
all the elements are there so that xu can start working with it without any problem, styling is not 100% ready, but this is not supposed to affect anything as xu is not supposed to deal with it at all. I will finish styling in the next couple of hours.

for the code cleaning, admin page, login pages, and index, article and collection pages, and my jijitang page are all code cleaned, I have added a new header for login page,

the first first pages are not yet cleaned as I have to delete a lot of files from there, not sure if I should do it now or leave it tmorrow until I confirm the pages we keep and the ones we delete,

another  commit will be followed before I arrive tomorrow to shanghai,